Move sessions filters logic to subfolder
