+++ frontend/src/scenes/sessions/filters/SessionsFilterBox.tsx
@@ -13,7 +13,7 @@ import { actionsModel } from '~/models/actionsModel'
 import { ActionType, CohortType } from '~/types'
 import { EntityTypes } from 'scenes/insights/trendsLogic'
 import { ActionInfo } from 'scenes/insights/ActionFilter/ActionFilterDropdown'
-import { sessionsFiltersLogic } from 'scenes/sessions/sessionsFiltersLogic'
+import { sessionsFiltersLogic } from 'scenes/sessions/filters/sessionsFiltersLogic'
 import { Link } from 'lib/components/Link'
 import { cohortsModel } from '~/models/cohortsModel'
 import { userLogic } from 'scenes/userLogic'
